Refactored according to Jimmy's comments as well as moving code currently common to Plista and Outbrain to be defined only once in the calling module.
